requirements.txt
# conda create -n SSTNet python=3.10.11

# conda install pytorch==1.12.0 torchvision==0.13.0 torchaudio==0.12.0 cudatoolkit=11.3 -c pytorch

    # torch==1.12.0
    # torchvision==0.13.0


# opencv-python==4.7.0.72
# scipy==1.10.1
# matplotlib
# tensorboard
# tqdm
# numpy==1.24.3
# pillow==9.5.0

opencv-python==4.7.0.72
scipy==1.10.1
matplotlib
tensorboard
tqdm
numpy==1.24.3
pillow==9.5.0
thop
pycocotools

# pip install -r requirements.txt